requirements.txt
requests
pypdf
Flask
Flask-CORS
openai
tavily-python
python-dotenv